include/prov: Rename to use ofi file name prefix
This is an internal header, add ofi prefix.
