+++ libfabric.vcxproj.filters
@@ -428,7 +428,7 @@
     <ClInclude Include="include\ofi_signal.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="include\prov.h">
+    <ClInclude Include="include\ofi_prov.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="include\rbtree.h">
