+++ prov/tcp/src/tcpx_comm.c
@@ -31,7 +31,7 @@
  */
 
 #include <rdma/fi_errno.h>
-#include <prov.h>
+#include <ofi_prov.h>
 #include <sys/types.h>
 #include <ofi_util.h>
 #include "tcpx.h"
