+++ prov/tcp/src/tcpx_init.c
@@ -32,7 +32,7 @@
 
 #include <rdma/fi_errno.h>
 
-#include <prov.h>
+#include <ofi_prov.h>
 #include "tcpx.h"
 
 #include <sys/types.h>
