+++ prov/udp/src/udpx_init.c
@@ -32,7 +32,7 @@
 
 #include <rdma/fi_errno.h>
 
-#include <prov.h>
+#include <ofi_prov.h>
 #include "udpx.h"
 
 #include <sys/types.h>
